feat: Open the setup wizard from Create workspace
Create workspace opened a dialog that asked again for the name that the
setup wizard collects. Users saw two create paths, and the new wizard
stayed hidden behind the old dialog.

Create workspace now opens the wizard at /workspaces/new. The wizard
creates the workspace when the user submits the name, so an abandoned
setup leaves no empty workspace. The server derives the workspace key,
and Settings keeps the key editable.

The setup screens now follow the onboarding design. Version control and
coding agent show the design cards, which open the real integration
dialog. The home installer and the wizard share one connect flow. The
workspace sidebar stays hidden until setup is complete.
