+++ web_src/src/pages/home/InstallIntegrationsSection.tsx
@@ -1,13 +1,6 @@
-import type { OrganizationsCreateIntegrationResponse, OrganizationsIntegration } from "@/api-client";
 import { Button } from "@/components/ui/button";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
-import { useMemo, useRef, useState, type MutableRefObject } from "react";
-import { useAvailableIntegrations, useConnectedIntegrations, useCreateIntegration } from "@/hooks/useIntegrations";
 import { IntegrationIcon } from "@/ui/componentSidebar/integrationIcons";
-import { IntegrationCreateDialog, type IntegrationCreatePayload } from "@/ui/IntegrationCreateDialog";
-import { ConfigureIntegrationDialog } from "@/ui/ConfigureIntegrationDialog";
-import { getIntegrationWebhookUrl } from "@/lib/integrationUtils";
-import { getNextIntegrationName } from "@/pages/organization/settings/components/IntegrationSetup/lib";
 import { getIntegrationTypeDisplayName } from "@/lib/integrationDisplayName";
 
 import { HomeIntegrationConnectRow, StatusDot } from "./HomeIntegrationConnectRow";
@@ -17,23 +10,9 @@ import {
   type IntegrationInstanceSummary,
   type IntegrationSelections,
 } from "./homeIntegrationStatus";
-import { useHomeIntegrationConnectActions } from "./useHomeIntegrationConnectActions";
-import { useInstallIntegrationSelections, useRefetchOnWindowFocus } from "./useInstallIntegrationSelections";
+import { useIntegrationConnectDialog } from "./useIntegrationConnectDialog";
 
 export type { IntegrationSelections };
-
-function resolveIntegrationHomeHref(args: {
-  organizationId: string;
-  dialogIntegrationName: string | null;
-  dialogMode: "create" | "resume";
-  pendingId?: string;
-  selectedId?: string;
-}) {
-  if (!args.organizationId) return undefined;
-  const integrationId = (args.dialogMode === "resume" ? args.pendingId : undefined) ?? args.selectedId;
-  if (integrationId) return `/${args.organizationId}/settings/integrations/${integrationId}`;
-  return `/${args.organizationId}/settings/integrations`;
-}
 
 export function IntegrationsSection({
   integrations,
@@ -49,205 +28,26 @@ export function IntegrationsSection({
   /** `status` shows Connected / Pending rows with switcher; `select` keeps the legacy instance picker. */
   variant?: "select" | "status";
 }) {
-  const { data: connected = [], refetch } = useConnectedIntegrations(organizationId, {
-    enabled: !!organizationId,
-  });
-  const { data: availableIntegrations = [] } = useAvailableIntegrations({ enabled: !!organizationId });
-  const createIntegrationMutation = useCreateIntegration(organizationId, "install_wizard");
-  const [dialogIntegrationName, setDialogIntegrationName] = useState<string | null>(null);
-  /** "create" skips resuming a pending instance so "Connect new" always starts fresh. */
-  const [dialogMode, setDialogMode] = useState<"create" | "resume">("resume");
-  const [configureIntegrationId, setConfigureIntegrationId] = useState<string | null>(null);
-  const pendingConnectKeyRef = useRef<string | null>(null);
-  const existingIntegrationNames = useMemo(
-    () => new Set(connected.map((i) => i.metadata?.name?.trim()).filter((n): n is string => Boolean(n))),
-    [connected],
-  );
-  const integrationData = useMemo(
-    () =>
-      integrations.map((name) => {
-        const allInstances = connected.filter((item) => item.metadata?.integrationName === name);
-        return { name, allInstances, readyInstances: allInstances.filter((item) => item.status?.state === "ready") };
-      }),
-    [integrations, connected],
-  );
-  const { rememberPreferredInstance } = useInstallIntegrationSelections({
-    integrationData,
+  const connect = useIntegrationConnectDialog({
+    organizationId,
+    integrationNames: integrations,
     selections,
     onSelectionsChange,
   });
-  useRefetchOnWindowFocus(refetch);
-
-  const { dialogDefinition, dialogPendingInstance, initialWebhookSetup, defaultDialogName } = useCreateDialogProps(
-    dialogIntegrationName,
-    availableIntegrations,
-    connected,
-    existingIntegrationNames,
-  );
-  const integrationHomeHref = useMemo(
-    () =>
-      resolveIntegrationHomeHref({
-        organizationId,
-        dialogIntegrationName,
-        dialogMode,
-        pendingId: dialogPendingInstance?.metadata?.id,
-        selectedId: dialogIntegrationName ? selections[dialogIntegrationName]?.id : undefined,
-      }),
-    [organizationId, dialogIntegrationName, dialogMode, dialogPendingInstance?.metadata?.id, selections],
-  );
-  const { openCapabilitySetup, openCreateIntegrationModal, openConnectDialog, openConfigureDialog } =
-    useHomeIntegrationConnectActions({
-      organizationId,
-      availableIntegrations,
-      connected,
-      pendingConnectKeyRef,
-      setDialogMode,
-      setDialogIntegrationName,
-      setConfigureIntegrationId,
-    });
-
-  const openCapabilitySetupAndPrefer = (integrationName: string, integrationId?: string) => {
-    if (integrationId) rememberPreferredInstance(integrationName, integrationId);
-    openCapabilitySetup(integrationName, integrationId);
-    void refetch();
-  };
 
   return (
     <>
       <IntegrationList
-        integrationData={integrationData}
+        integrationData={connect.integrationData}
         variant={variant}
         selections={selections}
         onSelectionsChange={onSelectionsChange}
-        onConnect={openConnectDialog}
-        onConfigure={openConfigureDialog}
-        onCreateNew={openCreateIntegrationModal}
+        onConnect={connect.requestConnect}
+        onConfigure={connect.configure}
+        onCreateNew={connect.createNew}
       />
-      <ConfigureIntegrationDialog
-        integrationId={configureIntegrationId}
-        organizationId={organizationId}
-        onClose={() => {
-          setConfigureIntegrationId(null);
-          void refetch();
-        }}
-      />
-      <HomeIntegrationCreateDialog
-        open={!!dialogIntegrationName}
-        dialogIntegrationName={dialogIntegrationName}
-        dialogMode={dialogMode}
-        organizationId={organizationId}
-        integrationHomeHref={integrationHomeHref}
-        dialogDefinition={dialogDefinition}
-        defaultDialogName={defaultDialogName}
-        existingIntegrationNames={existingIntegrationNames}
-        resumePendingForDialog={dialogMode === "resume" ? dialogPendingInstance : undefined}
-        initialWebhookSetup={initialWebhookSetup}
-        createIntegrationMutation={createIntegrationMutation}
-        pendingConnectKeyRef={pendingConnectKeyRef}
-        selections={selections}
-        onSelectionsChange={onSelectionsChange}
-        onPreferInstance={rememberPreferredInstance}
-        onClose={() => {
-          setDialogIntegrationName(null);
-          setDialogMode("resume");
-        }}
-        onCapabilitySetup={openCapabilitySetupAndPrefer}
-        onRefetch={() => void refetch()}
-      />
+      {connect.dialogs}
     </>
-  );
-}
-
-function HomeIntegrationCreateDialog({
-  open,
-  dialogIntegrationName,
-  dialogMode,
-  organizationId,
-  integrationHomeHref,
-  dialogDefinition,
-  defaultDialogName,
-  existingIntegrationNames,
-  resumePendingForDialog,
-  initialWebhookSetup,
-  createIntegrationMutation,
-  pendingConnectKeyRef,
-  selections,
-  onSelectionsChange,
-  onPreferInstance,
-  onClose,
-  onCapabilitySetup,
-  onRefetch,
-}: {
-  open: boolean;
-  dialogIntegrationName: string | null;
-  dialogMode: "create" | "resume";
-  organizationId: string;
-  integrationHomeHref?: string;
-  dialogDefinition: unknown;
-  defaultDialogName: string;
-  existingIntegrationNames: Set<string>;
-  resumePendingForDialog?: OrganizationsIntegration;
-  initialWebhookSetup?: { id: string; webhookUrl: string; config: Record<string, unknown> };
-  createIntegrationMutation: {
-    mutateAsync: (payload: IntegrationCreatePayload) => Promise<{ data: OrganizationsCreateIntegrationResponse }>;
-    reset: () => void;
-  };
-  pendingConnectKeyRef: MutableRefObject<string | null>;
-  selections: IntegrationSelections;
-  onSelectionsChange: (selections: IntegrationSelections) => void;
-  onPreferInstance: (integrationName: string, integrationId: string) => void;
-  onClose: () => void;
-  onCapabilitySetup: (integrationName: string, integrationId?: string) => void;
-  onRefetch: () => void;
-}) {
-  return (
-    <IntegrationCreateDialog
-      open={open}
-      onOpenChange={(nextOpen) => {
-        if (!nextOpen) {
-          onClose();
-          createIntegrationMutation.reset();
-        }
-      }}
-      integrationDefinition={(dialogDefinition as never) ?? null}
-      organizationId={organizationId}
-      integrationHomeHref={integrationHomeHref}
-      onCreateIntegration={async (payload) => {
-        const res = await createIntegrationMutation.mutateAsync(payload);
-        return res.data;
-      }}
-      onReset={() => createIntegrationMutation.reset()}
-      defaultName={
-        dialogMode === "create"
-          ? resolveDefaultDialogName(dialogIntegrationName, undefined, existingIntegrationNames)
-          : defaultDialogName
-      }
-      onCreated={(integrationId, instanceName) => {
-        // Dialog calls onOpenChange(false) before onCreated; keep the key in a ref across that close.
-        const key = pendingConnectKeyRef.current;
-        pendingConnectKeyRef.current = null;
-        if (key) {
-          onPreferInstance(key, integrationId);
-          // Newly created instances are not ready until setup finishes.
-          onSelectionsChange({
-            ...selections,
-            [key]: { id: integrationId, name: instanceName, ready: false },
-          });
-        }
-        onClose();
-        onRefetch();
-      }}
-      onCapabilitySetupRequired={(integrationName, integrationId) => {
-        pendingConnectKeyRef.current = null;
-        onPreferInstance(integrationName, integrationId);
-        onClose();
-        onCapabilitySetup(integrationName, integrationId);
-      }}
-      initialBrowserAction={resumePendingForDialog?.status?.browserAction}
-      initialCreatedIntegrationId={resumePendingForDialog?.metadata?.id}
-      initialWebhookSetup={dialogMode === "create" ? undefined : initialWebhookSetup}
-      initialConfiguration={resumePendingForDialog?.spec?.configuration as Record<string, unknown> | undefined}
-    />
   );
 }
 
@@ -311,51 +111,6 @@ function IntegrationList({
       )}
     </div>
   );
-}
-
-function buildWebhookSetup(pending: OrganizationsIntegration | undefined) {
-  const webhookUrl = getIntegrationWebhookUrl(pending?.status?.metadata);
-  if (!webhookUrl || !pending?.metadata?.id) return undefined;
-  return { id: pending.metadata.id, webhookUrl, config: { ...(pending.spec?.configuration ?? {}) } };
-}
-
-function resolveDefaultDialogName(
-  dialogIntegrationName: string | null,
-  pending: OrganizationsIntegration | undefined,
-  existingNames: Set<string>,
-): string {
-  if (pending?.metadata?.name) return pending.metadata.name;
-  if (!dialogIntegrationName) return "";
-  return getNextIntegrationName(dialogIntegrationName, existingNames);
-}
-
-function useCreateDialogProps(
-  dialogIntegrationName: string | null,
-  availableIntegrations: Array<{ name?: string; [key: string]: unknown }>,
-  connected: OrganizationsIntegration[],
-  existingIntegrationNames: Set<string>,
-) {
-  const dialogDefinition = useMemo(
-    () => (dialogIntegrationName ? availableIntegrations.find((d) => d.name === dialogIntegrationName) : undefined),
-    [availableIntegrations, dialogIntegrationName],
-  );
-
-  const dialogPendingInstance = useMemo(
-    () =>
-      dialogIntegrationName
-        ? connected.find((i) => i.metadata?.integrationName === dialogIntegrationName && i.status?.state !== "ready")
-        : undefined,
-    [dialogIntegrationName, connected],
-  );
-
-  const initialWebhookSetup = useMemo(() => buildWebhookSetup(dialogPendingInstance), [dialogPendingInstance]);
-
-  const defaultDialogName = useMemo(
-    () => resolveDefaultDialogName(dialogIntegrationName, dialogPendingInstance, existingIntegrationNames),
-    [dialogIntegrationName, dialogPendingInstance, existingIntegrationNames],
-  );
-
-  return { dialogDefinition, dialogPendingInstance, initialWebhookSetup, defaultDialogName };
 }
 
 function IntegrationRow({
